Got rid of all the linter complaints in HeaderView
