+++ client/.eslintrc.json
@@ -1,3 +1,7 @@
 {
-    "extends": "airbnb"
+    "extends": "airbnb",
+    "globals": {
+        "window": true,
+        "console": true
+      }
 }
