+++ client/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@okta/okta-signin-widget": "^2.3.0",
     "axios": "^0.17.0",
+    "prop-types": "^15.6.0",
     "react": "^16.0.0",
     "react-bootstrap": "^0.31.5",
     "react-bootstrap-sidebar": "0.0.1",
